All files cleaned and formatted. Compiled in windows and Linux no warnings.

+++ Connectivity_Matrix.cpp
@@ -17,21 +17,22 @@
 #include <stdexcept>
 #include "Connectivity_Matrix.hpp"
 
-Connectivity_Matrix::Connectivity_Matrix(const vector<double> &connections, const vector<string> &new_pages ): Page_Matrix(new_pages) {
+Connectivity_Matrix::Connectivity_Matrix( const vector<double> &connections, const vector<string> &new_pages )
+        : Page_Matrix( new_pages ) {
     double side_size = sqrt( connections.size());
 
-    if ( floor( side_size) != side_size) {
-        throw invalid_argument("Argument must have an integer square root");
+    if ( floor( side_size ) != side_size ) {
+        throw invalid_argument( "Argument must have an integer square root" );
     }
 
-    col_count = (int) side_size;
-    row_count = (int) side_size;
+    col_count = ( unsigned long ) side_size;
+    row_count = ( unsigned long ) side_size;
 
-    for (unsigned int i = 0; i < connections.size(); ++i) {
-        double val{connections[i]};
-        if (val != 0 && val != 1) {
-            throw invalid_argument("All elements in connections must have a value of 0 or 1");
+    for ( unsigned int i = 0; i < connections.size(); ++i ) {
+        double val{ connections[ i ] };
+        if ( val != 0 && val != 1 ) {
+            throw invalid_argument( "All elements in connections must have a value of 0 or 1" );
         }
-        matrix[i % (int)side_size][i / side_size] = val;
+        matrix[ i % ( int ) side_size ][ i / side_size ] = val;
     }
 }
